Split out aggregator implementation
